pyproject.toml
[project]
name = "timetensor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.9.4",
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "pynvml>=12.0.0",
    "scikit-learn>=1.6.0",
    "tabulate>=0.9.0",
    "torch>=2.5.1",
    "torchinfo>=1.8.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]